Minor changes to types and constrains in recommendation formation

+++ backend/DTO/GymRecomendationRequestDto.cs
@@ -10,8 +10,8 @@ public class GymRecommendationRequestDto
     public string? PreferredDepartureTime { get; set; } // Optional, format "hh:mm"
     public string? PreferredArrivalTime { get; set; } // Optional, format "hh:mm"
     public int MaxMembershipPrice { get; set; } // 0 to 10000 scale
-    public int MinOverallRating { get; set; } // 1 to 5 scale
-    public int MinCongestionRating { get; set; } // 1 to 5 scale
+    public decimal MinOverallRating { get; set; } // 1 to 5 scale
+    public decimal MinCongestionRating { get; set; } // 1 to 5 scale
     public double Latitude { get; set; } // Latitude of the location
     public double Longitude { get; set; } // Longitude of the location
 }
